requirements.txt
requests
beautifulsoup4
asyncio
python-telegram-bot
